refactor: Organizando melhor o codigo.
