+++ frontend/src/app/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import { AppSidebar } from "@/components/app-sidebar"
+import { AppSidebar } from "@/components/layout/app-sidebar"
 import {
   Breadcrumb,
   BreadcrumbItem,
@@ -13,7 +13,7 @@ import {
   SidebarProvider,
   SidebarTrigger,
 } from "@/components/ui/sidebar"
-import { DashboardContent } from "@/components/dashboard-content"
+import { DashboardContent } from "@/components/dashboard/dashboard-content"
 
 export default function Page() {
   return (
