+++ frontend/src/components/dashboard/dashboard-content.tsx
