+++ frontend/src/components/dashboard/dashboard-with-sidebar.tsx
