+++ frontend/src/components/dashboard/index.ts
@@ -1,0 +1,2 @@
+// Dashboard components
+export { DashboardContent } from './dashboard-content'
