+++ frontend/src/components/event-calendar/event-dialog.tsx
@@ -13,7 +13,7 @@ import {
   DefaultStartHour,
   EndHour,
   StartHour,
-} from "@/components/event-calendar/constants"
+} from "./constants"
 import { cn } from "@/lib/utils"
 import { Button } from "@/components/ui/button"
 import { Calendar } from "@/components/ui/calendar"
