+++ frontend/src/components/event-calendar/hooks/use-current-time-indicator.ts
@@ -6,7 +6,7 @@ import { endOfWeek, isSameDay, isWithinInterval, startOfWeek } from "date-fns"
 import {
   EndHour,
   StartHour,
-} from "@/components/event-calendar/constants"
+} from "../constants"
 
 export function useCurrentTimeIndicator(
   currentDate: Date,
