+++ frontend/src/components/event-calendar/week-view.tsx
@@ -30,7 +30,7 @@ import {
 import {
   EndHour,
   StartHour,
-} from "@/components/event-calendar/constants"
+} from "./constants"
 import { cn } from "@/lib/utils"
 
 interface WeekViewProps {
