+++ frontend/src/components/index.ts
@@ -1,0 +1,14 @@
+// Layout components
+export { AppSidebar } from './layout/app-sidebar'
+
+// Navigation components
+export { NavMain } from './navigation/nav-main'
+export { NavProjects } from './navigation/nav-projects'
+export { NavUser } from './navigation/nav-user'
+export { TeamSwitcher } from './navigation/team-switcher'
+
+// Dashboard components
+export { DashboardContent } from './dashboard/dashboard-content'
+
+// Event Calendar components
+export * from './event-calendar'
