+++ frontend/src/components/layout/index.ts
@@ -1,0 +1,2 @@
+// Layout components
+export { AppSidebar } from './app-sidebar'
