+++ frontend/src/components/navigation/index.ts
@@ -1,0 +1,5 @@
+// Navigation components
+export { NavMain } from './nav-main'
+export { NavProjects } from './nav-projects'
+export { NavUser } from './nav-user'
+export { TeamSwitcher } from './team-switcher'
