+++ frontend/src/components/navigation/nav-projects.tsx
