+++ frontend/src/components/navigation/team-switcher.tsx
